+++ packages/next/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/next
 
+## 4.3.7
+
+### Patch Changes
+
+- remove experimental .action output handling ([#11998](https://github.com/vercel/vercel/pull/11998))
+
 ## 4.3.6
 
 ### Patch Changes